Summary
add  flexibility
